refactor: move useMaskedErrors to core
