+++ packages/core/src/index.ts
@@ -6,3 +6,4 @@ export * from './plugins/use-schema';
 export * from './plugins/use-error-handler';
 export * from './plugins/use-extend-context';
 export * from './plugins/use-payload-formatter';
+export * from './plugins/use-masked-errors';
