--- packages/plugins/masked-errors/package.json
@@ -1,42 +1,0 @@
-{
-    "name": "@envelop/masked-errors",
-    "version": "0.1.0",
-    "author": "Dotan Simha <dotansimha@gmail.com>",
-    "license": "MIT",
-    "sideEffects": false,
-    "repository": {
-      "type": "git",
-      "url": "https://github.com/dotansimha/envelop.git",
-      "directory": "packages/plugins/graphql-modules"
-    },
-    "main": "dist/index.cjs.js",
-    "module": "dist/index.esm.js",
-    "typings": "dist/index.d.ts",
-    "typescript": {
-      "definition": "dist/index.d.ts"
-    },
-    "scripts": {
-      "test": "jest",
-      "prepack": "bob prepack"
-    },
-    "dependencies": {},
-    "devDependencies": {
-      "reflect-metadata": "0.1.13",
-      "graphql-modules": "1.4.2",
-      "bob-the-bundler": "1.2.0",
-      "graphql": "15.5.0",
-      "typescript": "4.2.4"
-    },
-    "peerDependencies": {
-      "graphql-modules": "^1",
-      "graphql": "^14.0.0 || ^15.0.0"
-    },
-    "buildOptions": {
-      "input": "./src/index.ts"
-    },
-    "publishConfig": {
-      "directory": "dist",
-      "access": "public"
-    }
-  }
-
